% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_it_up.R
\name{wrap_it_up}
\alias{wrap_it_up}
\title{Wrap your most used R packages in a given year}
\usage{
wrap_it_up(path, year = 2025)
}
\arguments{
\item{path}{path where your scripts can be found}

\item{year}{year you want to wrap}
}
\value{
top 5 functions and packages plus plot
}
\description{
This function scours all R scripts in a named path (.Rmd, .qmd, .R, etc.)
to determine your most-used packages for a specified year. Top packages
and top functions are printed and a plot showing your top packages is
generated.
}
