% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_library.R
\name{wrap_library}
\alias{wrap_library}
\title{Read in a Music xlm library.}
\usage{
wrap_library(path, year)
}
\arguments{
\item{path}{The path to your \code{Library.xml} file.}

\item{year}{The year to filter on, as an integer.}
}
\description{
Read in a Music xlm library.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
wrap_library("Library.xml", 2022L)
\dontshow{\}) # examplesIf}
}
