\name{Wrapped-package}
\alias{Wrapped-package}
\alias{Wrapped}
\docType{package}
\title{Computes Pdf, Cdf, Quantile, Random Numbers and Provides Estimation for 40 Univariate Wrapped Distributions}
\description{Computes the probability density function, cumulative distribution function, quantile function and random numbers for 40 univariate wrapped distributions.  They include the wrapped normal, wrapped Gumbel, wrapped logistic, wrapped t, wrapped Cauchy, wrapped skew normal, wrapped skew t, wrapped asymmetric Laplace, wrapped normal Laplace, wrapped skew Laplace, wrapped skew logistic, wrapped exponential power, wrapped skew power exponential, wrapped power exponential t, wrapped skew generalized t, wrapped skew hyperbolic, wrapped generalized hyperbolic Student t, wrapped power hyperbola logistic, wrapped Kiener, wrapped Laplace mixture, wrapped skew Laplace, wrapped polynomial tail Laplace, wrapped generalized asymmetric t, wrapped variance gamma, wrapped normal inverse gamma, wrapped skew Cauchy, wrapped slash, wrapped ex Gaussian, wrapped stable and wrapped log gamma distributions.  Also given are maximum likelihood estimates of the parameters, standard errors, 95 percent confidence intervals, log-likelihood values, AIC values, CAIC values, BIC values, HQIC values, values of the W statistic, values of the A statistic, values of the KS tatistic and the associated p-value.}
\details{
\tabular{ll}{
Package: \tab Wrapped\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-01-31\cr
License: \tab GPL(>=2)
}
probability density function, cumulative distribution function, quantile function, random numbers and measures of inference
}
\author{Saralees Nadarajah, Yuanyuan Zhang

Maintainer: Saralees Nadarajah <Saralees.Nadarajah@manchester.ac.uk>
}
\references{S. Nadarajah and Y. Zhang, Wrapped: A new R package for wrapped distributions, submitted.}
\keyword{package}
