% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped.R
\docType{data}
\name{wrapped}
\alias{wrapped}
\title{My listening data}
\format{
A data frame.
}
\source{
My Music app!
}
\usage{
wrapped
}
\description{
This data set is the (slightly modified) output of running
\code{\link[=wrap_library]{wrap_library()}} on my own library.
}
\keyword{datasets}
