\name{Wrapped-package}
\alias{Wrapped-package}
\alias{Wrapped}
\docType{package}
\title{Computes Pdf, Cdf, Quantile, Random Numbers and Provides Estimation for any Wrapped Distribution}
\description{Computes the probability density function, cumulative distribution function, quantile function and random numbers for any univariate wrapped distribution. Also computes maximum likelihood estimates of the parameters, standard errors, 95 percent confidence intervals, value of Cramer-von Misses statistic, value of Anderson Darling statistic, value of Kolmogorov Smirnov test statistic and its $p$-value, value of Akaike Information Criterion, value of Consistent Akaike Information Criterion, value of Bayesian Information Criterion, value of Hannan-Quinn information criterion, minimum value of the negative log-likelihood function and convergence status when some data are fitted by one of the following wrapped distributions: wrapped normal, wrapped Gumbel, wrapped logistic, wrapped Student's $t$, wrapped Cauchy, wrapped skew normal, wrapped skew $t$, wrapped skew Cauchy, wrapped asymmetric Laplace, wrapped normal Laplace, wrapped generalized logistic, wrapped skew Laplace, wrapped exponential power, wrapped  skew exponential type 1, wrapped skew exponential type 2, wrapped skew exponential type 3, wrapped skew exponential type 4, wrapped normal exponential $t$, wrapped skew normal type 2, wrapped ex Gaussian, wrapped skew $t$  type 1, wrapped skew $t$ type 3, wrapped skew $t$ type 4, wrapped skew $t$  type 5, wrapped sinh arcsinh, wrapped exponential generalized beta type 2, wrapped Johnson's $S_u$, wrapped skew generalized $t$, wrapped skew hyperbolic, wrapped asymmetric Laplace, wrapped polynomial tail Laplace, wrapped generalized asymmetric $t$, wrapped variance gamma, wrapped normal inverse Gaussian, wrapped hyperbolic, wrapped skew Laplace, wrapped slash, wrapped beta normal, wrapped Laplace, wrapped short tailed symmetric, wrapped log gamma distributions.}
\details{
\tabular{ll}{
Package: \tab Wrapped\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2017-08-21\cr
License: \tab GPL(>=2)
}
probability density function, cumulative distribution function, quantile function, random numbers and measures of inference
}
\author{Saralees Nadarajah, Yuanyuan Zhang

Maintainer: Saralees Nadarajah <Saralees.Nadarajah@manchester.ac.uk>
}
\references{S. Nadarajah and Y. Zhang, Wrapped: An R package for wrapped distributions, submitted.}
\keyword{package}
